requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
